refactor(list): change `ft_lstsize` return type to `size_t`
Change the return type of `ft_lstsize` from `int` to `size_t`.
This change prevents potential overflow on very large lists.
`size_t` is semantically correct for representing sizes and is consistent with standard C library functions.

+++ src/list/ft_lstsize.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int	ft_lstsize(t_list *lst)
+size_t	ft_lstsize(t_list *lst)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (lst)
